Moved HtmlTable in text_table
